requirements/local.txt
-r base.txt

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code quality
ruff>=0.4,<1.0
pyright>=1.1,<2.0
pre-commit>=3.0,<4.0

# Debugging
django-debug-toolbar>=4.3,<5.0
